agno-agent/pyproject.toml
[project]
name = "agno-agent"
version = "0.1.0"
description = "Agno agent with WebSocket interface for voice agent integration"
authors = [{name = "Nirmanus Team", email = "team@nirmanus.com"}]
dependencies = [
    "agno>=0.1.0",
    "fastapi>=0.100.0",
    "websockets>=12.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "mcp>=1.0.0",
    "ddgs>=9.9.1",
    "httpx>=0.27.0",
]

[project.scripts]
agno-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "pyproject.toml",
    "README.md",
    ".env.example"
]
